Fixing the c/ subdirectory size.c program and the .gitignore file there.
git-svn-id: svn+ssh://saulius-grazulis.lt/home/saulius/svn-repositories/compilers/trunk@3022 f7c8cbe2-f9e6-0310-9050-d7f47bf15973

+++ sc/c/sizes.c
@@ -3,9 +3,9 @@
 
 int main()
 {
-    printf( "sizeof(intmax_t) = %jd\n", sizeof(intmax_t) );
-    printf( "sizeof(ssize_t)  = %jd\n", sizeof(ssize_t) );
-    printf( "sizeof(void*)    = %jd\n", sizeof(void*) );
-    printf( "sizeof(int)      = %jd\n", sizeof(int) );
+    printf( "sizeof(intmax_t) = %jd\n", (intmax_t)sizeof(intmax_t) );
+    printf( "sizeof(ssize_t)  = %jd\n", (intmax_t)sizeof(ssize_t) );
+    printf( "sizeof(void*)    = %jd\n", (intmax_t)sizeof(void*) );
+    printf( "sizeof(int)      = %jd\n", (intmax_t)sizeof(int) );
     return 0;
 }
